// src/types/Characteristics.ts
export type Characteristics =
  | "Loves to eat"
  | "Proud of its power"
  | "Sturdy body"
  | "Likes to run"
  | "Highly curious"
  | "Strong willed"
  | "Takes plenty of siestas"
  | "Likes to thrash about"
  | "Capable of taking hits"
  | "Alert to sounds"
  | "Mischievous"
  | "Somewhat vain"
  | "Nods off a lot"
  | "A little quick tempered"
  | "Highly persistent"
  | "Impetuous and silly"
  | "Thoroughly cunning"
  | "Strongly defiant"
  | "Scatters things often"
  | "Likes to fight"
  | "Good endurance"
  | "Somewhat of a clown"
  | "Often lost in thought"
  | "Hates to lose"
  | "Likes to relax"
  | "Quick tempered"
  | "Good perseverance"
  | "Quick to flee"
  | "Very finicky"
  | "Somewhat stubborn";